+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 - Form values don't include empty number inputs and empty file inputs anymore
 
+### Changed
+- Renamed `input`, `change` and `submit` custom events to `inform-input`, `inform-change` and `inform-submit`.
+
 ## [1.4.0] - 2022-04-23
 ### Fixed
 - Calling `requestSubmit()` on the `<form>` was throwing an error